refactor: extract library crate and reorganize tests
Move core functionality from main.rs into lib.rs to enable use as a
library crate. Reorganize test structure into dedicated tests/ directory
with unit, integration, and CLI test modules.

- Add lib.rs with public API exports
- Update Cargo.toml with [lib] and [[bin]] sections
- Extract git helper functions in cache.rs
- Move tests to tests/ directory with proper module structure
