+++ tests/integration/mod.rs
@@ -1,0 +1,7 @@
+//! Integration tests for repoverlay commands.
+
+mod apply;
+mod create;
+mod remove;
+mod status;
+mod switch;
