+++ tests/integration_tests.rs
@@ -1,0 +1,4 @@
+//! Integration tests for repoverlay commands.
+
+mod common;
+mod integration;
