+++ tests/unit/mod.rs
@@ -1,0 +1,3 @@
+//! Unit tests for repoverlay.
+
+mod remove_section;
