+++ tests/unit_tests.rs
@@ -1,0 +1,3 @@
+//! Unit tests for repoverlay.
+
+mod unit;
